+++ src/main/java/com/github/alexthe666/rats/client/render/entity/LayerRatHeldItem.java
@@ -1,6 +1,5 @@
 package com.github.alexthe666.rats.client.render.entity;
 
-import com.github.alexthe666.rats.client.model.ModelPiratCannon;
 import com.github.alexthe666.rats.client.model.ModelRat;
 import com.github.alexthe666.rats.server.entity.EntityRat;
 import com.github.alexthe666.rats.server.items.RatsItemRegistry;
@@ -29,7 +28,7 @@ public class LayerRatHeldItem implements LayerRenderer<EntityRat> {
     }
 
     public void doRenderLayer(EntityRat entity, float limbSwing, float limbSwingAmount, float partialTicks, float ageInTicks, float netHeadYaw, float headPitch, float scale) {
-        if(!(this.renderer.getMainModel() instanceof  ModelRat)){
+        if (!(this.renderer.getMainModel() instanceof ModelRat)) {
             return;
         }
         ItemStack itemstack = entity.getHeldItem(EnumHand.MAIN_HAND);
@@ -45,7 +44,7 @@ public class LayerRatHeldItem implements LayerRenderer<EntityRat> {
             }
             Minecraft minecraft = Minecraft.getMinecraft();
             if (entity.holdInMouth && entity.getAnimation() != EntityRat.ANIMATION_EAT && entity.cookingProgress <= 0
-                    && !entity.hasUpgrade(RatsItemRegistry.RAT_UPGRADE_PLATTER)  && !entity.hasUpgrade( RatsItemRegistry.RAT_UPGRADE_LUMBERJACK) && !entity.hasUpgrade( RatsItemRegistry.RAT_UPGRADE_MINER)) {
+                    && !entity.hasUpgrade(RatsItemRegistry.RAT_UPGRADE_PLATTER) && !entity.hasUpgrade(RatsItemRegistry.RAT_UPGRADE_LUMBERJACK) && !entity.hasUpgrade(RatsItemRegistry.RAT_UPGRADE_MINER)) {
                 translateToHead();
                 GlStateManager.rotate(180.0F, 0.0F, 0.0F, 1.0F);
                 GlStateManager.rotate(180.0F, 0.0F, 1.0F, 0.0F);
@@ -57,23 +56,23 @@ public class LayerRatHeldItem implements LayerRenderer<EntityRat> {
                 GlStateManager.rotate(180.0F, 0.0F, 1.0F, 0.0F);
                 GlStateManager.rotate(20.0F, 1.0F, 0.0F, 0.0F);
                 GlStateManager.translate(-0.155F, -0.025, 0.125F);
-                if(entity.hasUpgrade(RatsItemRegistry.RAT_UPGRADE_PLATTER)){
+                if (entity.hasUpgrade(RatsItemRegistry.RAT_UPGRADE_PLATTER)) {
                     GlStateManager.translate(0F, 0.25F, 0F);
-                    if(itemstack.getItem() instanceof ItemBlock){
+                    if (itemstack.getItem() instanceof ItemBlock) {
                         GlStateManager.rotate(-90, 1.0F, 0.0F, 0.0F);
-                    }else{
+                    } else {
                         GlStateManager.translate(0F, -0.1F, -0.075F);
 
                     }
                 }
-                if(entity.hasUpgrade(RatsItemRegistry.RAT_UPGRADE_LUMBERJACK) || entity.hasUpgrade(RatsItemRegistry.RAT_UPGRADE_MINER)) {
+                if (entity.hasUpgrade(RatsItemRegistry.RAT_UPGRADE_LUMBERJACK) || entity.hasUpgrade(RatsItemRegistry.RAT_UPGRADE_MINER)) {
                     GlStateManager.translate(0.15F, -0.075F, 0);
                 }
             }
             minecraft.getItemRenderer().renderItem(entity, itemstack, ItemCameraTransforms.TransformType.GROUND);
             GlStateManager.popMatrix();
         }
-        if(entity.hasUpgrade(RatsItemRegistry.RAT_UPGRADE_BUCCANEER)) {
+        if (entity.hasUpgrade(RatsItemRegistry.RAT_UPGRADE_BUCCANEER)) {
             GlStateManager.pushMatrix();
             ((ModelRat) this.renderer.getMainModel()).body1.postRender(0.0625F);
             GlStateManager.pushMatrix();
@@ -83,7 +82,7 @@ public class LayerRatHeldItem implements LayerRenderer<EntityRat> {
             this.renderer.bindTexture(LayerPiratBoatSail.TEXTURE_PIRATE_CANNON);
             LayerPiratBoatSail.MODEL_PIRAT_CANNON.render(entity, limbSwing, limbSwingAmount, ageInTicks, netHeadYaw, headPitch, scale);
             GlStateManager.popMatrix();
-            if(entity.getVisualFlag()){
+            if (entity.getVisualFlag()) {
                 GlStateManager.pushMatrix();
                 GlStateManager.translate(0, -0.925F, 0.2F);
                 GlStateManager.scale(0.5F, 0.5F, 0.5F);
@@ -97,7 +96,7 @@ public class LayerRatHeldItem implements LayerRenderer<EntityRat> {
             GlStateManager.popMatrix();
             GlStateManager.popMatrix();
         }
-        if(entity.hasUpgrade(RatsItemRegistry.RAT_UPGRADE_PLATTER)){
+        if (entity.hasUpgrade(RatsItemRegistry.RAT_UPGRADE_PLATTER)) {
             if (this.renderer.getMainModel().isChild) {
                 GlStateManager.translate(0.0F, 0.625F, 0.0F);
                 GlStateManager.rotate(-20.0F, -1.0F, 0.0F, 0.0F);
@@ -113,7 +112,7 @@ public class LayerRatHeldItem implements LayerRenderer<EntityRat> {
             GlStateManager.scale(2F, 2F, 2F);
             minecraft.getItemRenderer().renderItem(entity, PLATTER_STACK, ItemCameraTransforms.TransformType.GROUND);
         }
-        if(entity.hasUpgrade(RatsItemRegistry.RAT_UPGRADE_CRAFTING)){
+        if (entity.hasUpgrade(RatsItemRegistry.RAT_UPGRADE_CRAFTING)) {
             Minecraft minecraft = Minecraft.getMinecraft();
             GlStateManager.pushMatrix();
             translateToHand(true);
@@ -130,7 +129,7 @@ public class LayerRatHeldItem implements LayerRenderer<EntityRat> {
             minecraft.getItemRenderer().renderItem(entity, PICKAXE_STACK, ItemCameraTransforms.TransformType.GROUND);
             GlStateManager.popMatrix();
         }
-        if(entity.hasUpgrade(RatsItemRegistry.RAT_UPGRADE_LUMBERJACK)){
+        if (entity.hasUpgrade(RatsItemRegistry.RAT_UPGRADE_LUMBERJACK)) {
             Minecraft minecraft = Minecraft.getMinecraft();
             GlStateManager.pushMatrix();
             translateToHand(false);
@@ -140,7 +139,7 @@ public class LayerRatHeldItem implements LayerRenderer<EntityRat> {
             minecraft.getItemRenderer().renderItem(entity, IRON_AXE_STACK, ItemCameraTransforms.TransformType.GROUND);
             GlStateManager.popMatrix();
         }
-        if(entity.hasUpgrade(RatsItemRegistry.RAT_UPGRADE_MINER)){
+        if (entity.hasUpgrade(RatsItemRegistry.RAT_UPGRADE_MINER)) {
             Minecraft minecraft = Minecraft.getMinecraft();
             GlStateManager.pushMatrix();
             translateToHand(false);
@@ -149,9 +148,9 @@ public class LayerRatHeldItem implements LayerRenderer<EntityRat> {
             minecraft.getItemRenderer().renderItem(entity, new ItemStack(Items.DIAMOND_PICKAXE), ItemCameraTransforms.TransformType.GROUND);
             GlStateManager.popMatrix();
         }
-        if(entity.hasUpgrade(RatsItemRegistry.RAT_UPGRADE_FLIGHT)) {
+        if (entity.hasUpgrade(RatsItemRegistry.RAT_UPGRADE_FLIGHT)) {
             Minecraft minecraft = Minecraft.getMinecraft();
-            float wingAngle = entity.onGround ? 0 : (float)MathHelper.sin(ageInTicks) * 30;
+            float wingAngle = entity.onGround ? 0 : MathHelper.sin(ageInTicks) * 30;
             float wingFold = entity.onGround ? -45 : 0;
             ((ModelRat) this.renderer.getMainModel()).body1.postRender(0.0625F);
             ((ModelRat) this.renderer.getMainModel()).body2.postRender(0.0625F);
@@ -187,10 +186,10 @@ public class LayerRatHeldItem implements LayerRenderer<EntityRat> {
     protected void translateToHand(boolean left) {
         ((ModelRat) this.renderer.getMainModel()).body1.postRender(0.0625F);
         ((ModelRat) this.renderer.getMainModel()).body2.postRender(0.0625F);
-        if(left){
+        if (left) {
             ((ModelRat) this.renderer.getMainModel()).leftArm.postRender(0.0625F);
             ((ModelRat) this.renderer.getMainModel()).leftHand.postRender(0.0625F);
-        }else{
+        } else {
             ((ModelRat) this.renderer.getMainModel()).rightArm.postRender(0.0625F);
             ((ModelRat) this.renderer.getMainModel()).rightHand.postRender(0.0625F);
         }
