+++ src/main/java/com/github/alexthe666/rats/server/entity/ai/FlyingRatPathNavigate.java
@@ -3,7 +3,6 @@ package com.github.alexthe666.rats.server.entity.ai;
 import com.github.alexthe666.rats.server.entity.EntityRat;
 import com.github.alexthe666.rats.server.items.RatsItemRegistry;
 import net.minecraft.entity.Entity;
-import net.minecraft.entity.EntityLiving;
 import net.minecraft.pathfinding.Path;
 import net.minecraft.pathfinding.PathNavigateFlying;
 import net.minecraft.world.World;
@@ -19,7 +18,7 @@ public class FlyingRatPathNavigate extends PathNavigateFlying {
 
     public boolean tryMoveToEntityLiving(Entity entityIn, double speedIn) {
         if (rat.hasUpgrade(RatsItemRegistry.RAT_UPGRADE_FLIGHT)) {
-            rat.getMoveHelper().setMoveTo((double) entityIn.posX, (double) entityIn.posY + entityIn.height, (double) entityIn.posZ, 0.25D);
+            rat.getMoveHelper().setMoveTo(entityIn.posX, entityIn.posY + entityIn.height, entityIn.posZ, 0.25D);
         }
         Path path = this.getPathToEntityLiving(entityIn);
         return path != null && this.setPath(path, speedIn);
